requirements.txt
aiohttp
python-dotenv
loguru
beautifulsoup4
